:triangular_ruler: Move admin views into the views file
